requirements.txt
python-telegram-bot==10.1.0
PySocks==1.6.8